+++ core/olap/lib/util/functions.hpp
@@ -131,7 +131,7 @@ extern "C" void flushInt(int toFlush, char *fileName);
 
 extern "C" void flushDString(int toFlush, void *dict, char *fileName);
 
-extern "C" void flushInt64(size_t toFlush, char *fileName);
+extern "C" void flushInt64(int64_t toFlush, char *fileName);
 
 extern "C" void flushDate(int64_t toFlush, char *fileName);
